draft, moving create_map func to graphics
